rename: 파일명 수정 (globalUtils -> formatNumberWithCommas)
